Enhancement to properly give output of negative numbers
This includes enhancement to output a negative number when it
receives single negative number or expression to solve in the
input.

As per the previous logic the second child of the node was
always assumed as an operator, hence the scenario
where it receives a single negative number/expression was
parsed as a symbolic subtraction operator followed by
a expression or a number. However, in such cases the
answer of whatever comes from parsing the further expression
or number must be negated once at the end, which was missing.
This commit now covers this.