Enhance `zval` `zlist` constructor function : `zval_zlist()`
Enhanced `zval` `zlist` constructor function [ `zval_zlist()` ] .
Rather than just creating empty `zlists` this could now create
`zlists` of the specified size. Empty `zlists` could be created
by keeping the `count` to 0 . As `count` could never be a negative
an empty `zlist` is created if the constructor is provided with
a negative `count`. However, it is best to avoid such usage.