Include functions to create and delete zuzeelik environment
Included functions to create and delete zuzeelik environment
(`zenv`). This includes a new constructor function called
`zenv_create()` for constructing zenvs and a delete function
called `zenv_delete()` to delete zenvs i.e. to iterate over the
items in both the lists of zenv and free them.

This also includes macros such as `ZENV_COUNT` which is used to
get the total count of items in the lists of `zenv` and other
macros such as ZENV_SYM_LIST and ZENV_VAL_LIST to access the
pointers of the two lists.