Limit usage of macros related to zuzeelik environment (`zenv`)
Minimized the usage of macros related to zuzeelik environment
(`zenv`) . As the structures/unions of `zenv` is not that
complex as compared to that of `zval`s there is no need to
use macros for accessing its members. On the contrary,
it is more obtrusive if we use macros to access the members
of `zenv` structure.

Still, definitions of those macros are kept in case
it comes useful in later developments.