Add functions to get values from `zenv` or put values into it
Included functions to get values from zuzeelik environment
(`zenv`) or put values into it.

This includes two functions namely `zval_retrieve()` which
gets values from the environment and `zval_store()` which puts
values into the environment.

To get the values from the environment a loop is done over all the
items in the environment and it checks if any given symbols matches
any of the stored strings. If a match is found a copy is returned
of the stored value, if no value is found an error is returned.

To put values in the environment, first, a check is done to verify
if a variable with same name already exists. If it does, its value
is replaced with the new one. If no existing value is found with
the name then some space is allocated before it is put in.